Moved EditRegistrationPeriodsView to components
