+++ frontend/shared/components/src/periods/RegistrationPeriodRow.vue
