+++ frontend/shared/components/src/periods/index.ts
@@ -1,0 +1,1 @@
+export { default as EditRegistrationPeriodsView } from './EditRegistrationPeriodsView.vue';
